Introduce useBalance in favor of selectors
